+++ packages/trigger-sdk/src/v3/otel.ts
@@ -1,7 +1,9 @@
+import { metrics } from "@opentelemetry/api";
 import { traceContext } from "@trigger.dev/core/v3";
 
 export const otel = {
   withExternalTrace: <T>(fn: () => T): T => {
     return traceContext.withExternalTrace(fn);
   },
+  metrics,
 };